refactor: remove ActorCardList from sharable components
