+++ src/pages/MoviePage/ActorCardList.jsx
@@ -1,9 +1,12 @@
 import { Grid } from "@mui/material";
-import ActorCard from "../ActorCard/ActorCard";
+import ActorCard from "../../components/ActorCard/ActorCard";
+import { credits } from "../../mock-data/mockMovieDetail";
 
-const ActorCardList = ({ cast }) => {
+const ActorCardList = () => {
+  const cast = credits.cast?.slice(0, 5);
+
   return (
-    <Grid container spacing={3}>
+    <Grid container spacing={3} sx={{ mt: 2 }}>
       {cast &&
         cast.map((card) => (
           <Grid item key={card.id}>
